+++ rules/implementation/config.js
@@ -5,10 +5,13 @@ module.exports = {
   taskDefinition: {
     'bpmn:BusinessRuleTask': '1.1',
     'bpmn:IntermediateThrowEvent': {
-      'bpmn:MessageEventDefinition': '1.2',
+      'bpmn:MessageEventDefinition': '1.2'
     },
     'bpmn:ScriptTask': '1.1',
     'bpmn:SendTask': '1.1',
     'bpmn:ServiceTask': '1.0'
+  },
+  script: {
+    'bpmn:ScriptTask': '8.2'
   }
 };